+++ integration-tests/uhppoted/tcp.py
@@ -468,6 +468,27 @@ class TestUhppoteWithTCP(unittest.TestCase):
 
         self.assertEqual(response, SetDoorPasscodesResponse)
 
+    def test_get_antipassback(self):
+        '''
+        Tests the get_antipassback function with defaults.
+        '''
+        controller = (CONTROLLER, DEST_ADDR, 'tcp')
+
+        response = self.u.get_antipassback(controller)
+
+        self.assertEqual(response, GetAntiPassbackResponse)
+
+    def test_set_antipassback(self):
+        '''
+        Tests the set_antipassback function with defaults.
+        '''
+        controller = (CONTROLLER, DEST_ADDR, 'tcp')
+        antipassback = 2
+
+        response = self.u.set_antipassback(controller, antipassback)
+
+        self.assertEqual(response, SetAntiPassbackResponse)
+
     def test_restore_default_parameters(self):
         '''
         Tests the restore-default-parameters function with defaults.