+++ integration-tests/uhppoted/udp_default.py
@@ -474,6 +474,25 @@ class TestUDPWithDestAddr(unittest.TestCase):
 
         self.assertEqual(response, SetDoorPasscodesResponse)
 
+    def test_get_antipassback(self):
+        '''
+        Tests the get_antipassback function with defaults.
+        '''
+        controller = CONTROLLER
+        response = self.u.get_antipassback(controller)
+
+        self.assertEqual(response, GetAntiPassbackResponse)
+
+    def test_set_antipassback(self):
+        '''
+        Tests the set_antipassback function with defaults.
+        '''
+        controller = CONTROLLER
+        antipassback = 2
+        response = self.u.set_antipassback(controller, antipassback)
+
+        self.assertEqual(response, SetAntiPassbackResponse)
+
     def test_restore_default_parameters(self):
         '''
         Tests the restore-default-parameters function with defaults.